requirements.txt
asgiref==3.2.8
Django==3.0.7
django-cors-headers==3.4.0
django-summernote==0.8.11.6
djangorestframework==3.11.0
Pillow
psycopg2
psycopg2-binary
pytz==2020.1
sqlparse==0.3.1
